add kill.sh and existing program check

+++ kill.sh
@@ -1,5 +1,16 @@
-PIDS=`ps aux | grep -E "[0-9]{2} bf_switchd.{10,}nat" | sed "s/[^0-9]*\([0-9]*\).*/\1/g"`
-if [ -n "$PIDS" ] # not empty
+#!/bin/bash
+
+if [ $# -ne 1 ];
+then 
+	echo "Usage: kill.sh <P4_PROGRAM_NAME>"
+	exit 1
+fi
+
+PROGRAM="$1"
+PIDS=`pgrep bf_switchd`
+MATCH=`ps aux | grep -E "$PIDS.{6,}bf_switchd.+$PROGRAM.conf"`
+
+if [ -n "$PIDS" ] && [ -n "$MATCH" ] # not empty
 then
 	echo "These running switch processes will be killed: "$PIDS
 	kill -9 $PIDS > /dev/null 2>&1 #This will cause an error because it also kill the process of "grep", however it has terminated.
